requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
pandas==2.0.3
kociemba==1.2.1
streamlit==1.28.1
flask==2.3.3
flask-cors==4.0.0
watchdog==3.0.0 